Make BeginFrameSource source_id_ uint64_t.
In order to include part of the source id that changes when the GPU
process restarts, make the BeginFrameSource 64 bits larger. This CL
updates all usage of uint32_t for source_id to uint64_t along with some
constants, ParamTraits and StructTraits.

The only issue is that source_id is used to build TraceValue which only
supports int. The lower 32 bits are the interesting part so continue to
output only them. Trace output should be identical as a result.

Bug: 782268
Cq-Include-Trybots: master.tryserver.blink:linux_trusty_blink_rel

+++ components/viz/common/frame_sinks/begin_frame_args.cc
@@ -35,7 +35,7 @@ BeginFrameArgs::BeginFrameArgs()
       type(BeginFrameArgs::INVALID),
       on_critical_path(true) {}
 
-BeginFrameArgs::BeginFrameArgs(uint32_t source_id,
+BeginFrameArgs::BeginFrameArgs(uint64_t source_id,
                                uint64_t sequence_number,
                                base::TimeTicks frame_time,
                                base::TimeTicks deadline,
@@ -52,7 +52,7 @@ BeginFrameArgs::BeginFrameArgs(uint32_t source_id,
 }
 
 BeginFrameArgs BeginFrameArgs::Create(BeginFrameArgs::CreationLocation location,
-                                      uint32_t source_id,
+                                      uint64_t source_id,
                                       uint64_t sequence_number,
                                       base::TimeTicks frame_time,
                                       base::TimeTicks deadline,
@@ -112,7 +112,7 @@ BeginFrameAck::BeginFrameAck()
       source_id(0),
       has_damage(false) {}
 
-BeginFrameAck::BeginFrameAck(uint32_t source_id,
+BeginFrameAck::BeginFrameAck(uint64_t source_id,
                              uint64_t sequence_number,
                              bool has_damage)
     : sequence_number(sequence_number),
